refactored handover packages for records, encoder and decoders
