+++ nfctools-ndef/src/main/java/org/nfctools/ndef/NdefContext.java
@@ -31,12 +31,6 @@ import org.nfctools.ndef.wkt.decoder.GenericControlRecordDecoder;
 import org.nfctools.ndef.wkt.decoder.SmartPosterRecordDecoder;
 import org.nfctools.ndef.wkt.decoder.TextRecordDecoder;
 import org.nfctools.ndef.wkt.decoder.UriRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.AlternativeCarrierRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.CollisionResolutionRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.ErrorRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.HandoverCarrierRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.HandoverRequestRecordDecoder;
-import org.nfctools.ndef.wkt.decoder.handover.HandoverSelectRecordDecoder;
 import org.nfctools.ndef.wkt.encoder.ActionRecordEncoder;
 import org.nfctools.ndef.wkt.encoder.GcActionRecordEncoder;
 import org.nfctools.ndef.wkt.encoder.GcDataRecordEncoder;
@@ -45,12 +39,24 @@ import org.nfctools.ndef.wkt.encoder.GenericControlRecordEncoder;
 import org.nfctools.ndef.wkt.encoder.SmartPosterRecordEncoder;
 import org.nfctools.ndef.wkt.encoder.TextRecordEncoder;
 import org.nfctools.ndef.wkt.encoder.UriRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.AlternativeCarrierRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.CollisionResolutionRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.ErrorRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.HandoverCarrierRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.HandoverRequestRecordEncoder;
-import org.nfctools.ndef.wkt.encoder.handover.HandoverSelectRecordEncoder;
+import org.nfctools.ndef.wkt.handover.decoder.AlternativeCarrierRecordDecoder;
+import org.nfctools.ndef.wkt.handover.decoder.CollisionResolutionRecordDecoder;
+import org.nfctools.ndef.wkt.handover.decoder.ErrorRecordDecoder;
+import org.nfctools.ndef.wkt.handover.decoder.HandoverCarrierRecordDecoder;
+import org.nfctools.ndef.wkt.handover.decoder.HandoverRequestRecordDecoder;
+import org.nfctools.ndef.wkt.handover.decoder.HandoverSelectRecordDecoder;
+import org.nfctools.ndef.wkt.handover.encoder.AlternativeCarrierRecordEncoder;
+import org.nfctools.ndef.wkt.handover.encoder.CollisionResolutionRecordEncoder;
+import org.nfctools.ndef.wkt.handover.encoder.ErrorRecordEncoder;
+import org.nfctools.ndef.wkt.handover.encoder.HandoverCarrierRecordEncoder;
+import org.nfctools.ndef.wkt.handover.encoder.HandoverRequestRecordEncoder;
+import org.nfctools.ndef.wkt.handover.encoder.HandoverSelectRecordEncoder;
+import org.nfctools.ndef.wkt.handover.records.AlternativeCarrierRecord;
+import org.nfctools.ndef.wkt.handover.records.CollisionResolutionRecord;
+import org.nfctools.ndef.wkt.handover.records.ErrorRecord;
+import org.nfctools.ndef.wkt.handover.records.HandoverCarrierRecord;
+import org.nfctools.ndef.wkt.handover.records.HandoverRequestRecord;
+import org.nfctools.ndef.wkt.handover.records.HandoverSelectRecord;
 import org.nfctools.ndef.wkt.records.ActionRecord;
 import org.nfctools.ndef.wkt.records.GcActionRecord;
 import org.nfctools.ndef.wkt.records.GcDataRecord;
@@ -60,12 +66,6 @@ import org.nfctools.ndef.wkt.records.SmartPosterRecord;
 import org.nfctools.ndef.wkt.records.TextRecord;
 import org.nfctools.ndef.wkt.records.UriRecord;
 import org.nfctools.ndef.wkt.records.WellKnownRecord;
-import org.nfctools.ndef.wkt.records.handover.AlternativeCarrierRecord;
-import org.nfctools.ndef.wkt.records.handover.CollisionResolutionRecord;
-import org.nfctools.ndef.wkt.records.handover.ErrorRecord;
-import org.nfctools.ndef.wkt.records.handover.HandoverCarrierRecord;
-import org.nfctools.ndef.wkt.records.handover.HandoverRequestRecord;
-import org.nfctools.ndef.wkt.records.handover.HandoverSelectRecord;
 
 /**
  * Simple entry class for getting the different encoders and decoders.
