+++ nfctools-ndef/src/main/java/org/nfctools/ndef/wkt/handover/decoder/CollisionResolutionRecordDecoder.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.nfctools.ndef.wkt.decoder.handover;
+package org.nfctools.ndef.wkt.handover.decoder;
 
 import org.nfctools.ndef.NdefMessageDecoder;
 import org.nfctools.ndef.wkt.WellKnownRecordPayloadDecoder;
+import org.nfctools.ndef.wkt.handover.records.CollisionResolutionRecord;
 import org.nfctools.ndef.wkt.records.WellKnownRecord;
-import org.nfctools.ndef.wkt.records.handover.CollisionResolutionRecord;
 
 /**
  *
