+++ nfctools-ndef/src/main/java/org/nfctools/ndef/wkt/handover/encoder/ErrorRecordEncoder.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.nfctools.ndef.wkt.decoder.handover;
+package org.nfctools.ndef.wkt.handover.encoder;
 
-import org.nfctools.ndef.NdefMessageDecoder;
-import org.nfctools.ndef.wkt.WellKnownRecordPayloadDecoder;
+import org.nfctools.ndef.NdefMessageEncoder;
+import org.nfctools.ndef.wkt.WellKnownRecordPayloadEncoder;
+import org.nfctools.ndef.wkt.handover.records.ErrorRecord;
+import org.nfctools.ndef.wkt.handover.records.ErrorRecord.ErrorReason;
 import org.nfctools.ndef.wkt.records.WellKnownRecord;
-import org.nfctools.ndef.wkt.records.handover.ErrorRecord;
-import org.nfctools.ndef.wkt.records.handover.ErrorRecord.ErrorReason;
 
 /**
  * 
@@ -28,17 +28,25 @@ import org.nfctools.ndef.wkt.records.handover.ErrorRecord.ErrorReason;
  * 
  */
 
-public class ErrorRecordDecoder implements WellKnownRecordPayloadDecoder {
+public class ErrorRecordEncoder implements WellKnownRecordPayloadEncoder {
 
 	@Override
-	public WellKnownRecord decodePayload(byte[] payload, NdefMessageDecoder messageDecoder) {
-		ErrorRecord errorRecord = new ErrorRecord();
+	public byte[] encodePayload(WellKnownRecord wellKnownRecord, NdefMessageEncoder messageEncoder) {
 
-		ErrorReason errorReason = ErrorReason.toErrorReason(payload[0]);
+		ErrorRecord errorRecord = (ErrorRecord)wellKnownRecord;
 
-		errorRecord.setErrorReason(errorReason);
+		if (!errorRecord.hasErrorReason()) {
+			throw new IllegalArgumentException("Expected error reason");
+		}
 
-		Number number;
+		if (!errorRecord.hasErrorData()) {
+			throw new IllegalArgumentException("Expected error data");
+		}
+
+		ErrorReason errorReason = errorRecord.getErrorReason();
+
+		byte[] payload;
+
 		switch (errorReason) {
 			case TemporaryMemoryConstraints: {
 				/**
@@ -48,7 +56,7 @@ public class ErrorRecordDecoder implements WellKnownRecordPayloadDecoder {
 				 * the subsequent receipt of a Handover Request Message by the Handover Selector.
 				 */
 
-				number = new Short((short)(payload[1] & 0xFFFF));
+				payload = new byte[] { errorReason.getValue(), (byte)(errorRecord.getErrorData().shortValue() & 0xFF) };
 
 				break;
 			}
@@ -59,9 +67,10 @@ public class ErrorRecordDecoder implements WellKnownRecordPayloadDecoder {
 				 * number of octets of an acceptable Handover Select Message. The number of octets SHALL be determined
 				 * by the total length of the NDEF message, including all header information.
 				 */
-
-				number = new Long(((long)(payload[1] & 0xFF) << 24) + ((payload[2] & 0xFF) << 16)
-						+ ((payload[3] & 0xFF) << 8) + ((payload[4] & 0xFF) << 0));
+				long unsignedInt = errorRecord.getErrorData().longValue();
+				payload = new byte[] { errorReason.getValue(), (byte)((unsignedInt >> 24) & 0xFF),
+						(byte)((unsignedInt >> 16) & 0xFF), (byte)((unsignedInt >> 8) & 0xFF),
+						(byte)(unsignedInt & 0xFF) };
 
 				break;
 			}
@@ -74,7 +83,7 @@ public class ErrorRecordDecoder implements WellKnownRecordPayloadDecoder {
 				 * Handover Request Message by the Handover Selector.
 				 */
 
-				number = new Short((short)(payload[1] & 0xFFFF));
+				payload = new byte[] { errorReason.getValue(), (byte)(errorRecord.getErrorData().shortValue() & 0xFF) };
 
 				break;
 			}
@@ -83,9 +92,7 @@ public class ErrorRecordDecoder implements WellKnownRecordPayloadDecoder {
 			}
 		}
 
-		errorRecord.setErrorData(number);
-
-		return errorRecord;
+		return payload;
 	}
 
 }
